Added a LoginForm and refactored code in AddToList
Added a login form in order to implement a login in the future. Backend
is more or less finished. How is the JW Token handled in the frontend?

Added a ternary operator for the numberOfWheels drop down menu. If a car
is selected, the drop down menu appears. Also, moved the select options
outside of the return statement so that the code is easier to read.